+++ internal/driver/glfw/window.go
@@ -26,12 +26,12 @@ const (
 )
 
 var (
-	cursorMap    map[desktop.Cursor]*glfw.Cursor
+	cursorMap    map[desktop.StandardCursor]*glfw.Cursor
 	defaultTitle = "Fyne Application"
 )
 
 func initCursors() {
-	cursorMap = map[desktop.Cursor]*glfw.Cursor{
+	cursorMap = map[desktop.StandardCursor]*glfw.Cursor{
 		desktop.DefaultCursor:   glfw.CreateStandardCursor(glfw.ArrowCursor),
 		desktop.TextCursor:      glfw.CreateStandardCursor(glfw.IBeamCursor),
 		desktop.CrosshairCursor: glfw.CreateStandardCursor(glfw.CrosshairCursor),
@@ -52,11 +52,12 @@ type window struct {
 	decorate   bool
 	fixedSize  bool
 
-	cursor   *glfw.Cursor
-	canvas   *glCanvas
-	title    string
-	icon     fyne.Resource
-	mainmenu *fyne.MainMenu
+	cursor       desktop.Cursor
+	customCursor *glfw.Cursor
+	canvas       *glCanvas
+	title        string
+	icon         fyne.Resource
+	mainmenu     *fyne.MainMenu
 
 	clipboard fyne.Clipboard
 
@@ -558,22 +559,31 @@ func (w *window) findObjectAtPositionMatching(canvas *glCanvas, mouse fyne.Posit
 	return driver.FindObjectAtPositionMatching(mouse, matches, canvas.Overlays().Top(), canvas.menu, canvas.Content())
 }
 
-func fyneToNativeCursor(cursor desktop.Cursor) *glfw.Cursor {
-	ret, ok := cursorMap[cursor]
-	if !ok {
-		return cursorMap[desktop.DefaultCursor]
+func fyneToNativeCursor(cursor desktop.Cursor) (*glfw.Cursor, bool) {
+	switch v := cursor.(type) {
+	case desktop.StandardCursor:
+		ret, ok := cursorMap[v]
+		if !ok {
+			return cursorMap[desktop.DefaultCursor], false
+		}
+		return ret, false
+	default:
+		img, x, y := cursor.Image()
+		if img == nil {
+			return nil, true
+		}
+		return glfw.CreateCursor(img, x, y), true
 	}
-	return ret
 }
 
 func (w *window) mouseMoved(viewport *glfw.Window, xpos float64, ypos float64) {
 	w.mousePos = fyne.NewPos(internal.UnscaleInt(w.canvas, int(xpos)), internal.UnscaleInt(w.canvas, int(ypos)))
 
-	cursor := cursorMap[desktop.DefaultCursor]
+	cursor := desktop.Cursor(desktop.DefaultCursor)
+
 	obj, pos, _ := w.findObjectAtPositionMatching(w.canvas, w.mousePos, func(object fyne.CanvasObject) bool {
 		if cursorable, ok := object.(desktop.Cursorable); ok {
-			fyneCursor := cursorable.Cursor()
-			cursor = fyneToNativeCursor(fyneCursor)
+			cursor = cursorable.Cursor()
 		}
 
 		_, hover := object.(desktop.Hoverable)
@@ -582,12 +592,21 @@ func (w *window) mouseMoved(viewport *glfw.Window, xpos float64, ypos float64) {
 
 	if w.cursor != cursor {
 		// cursor has changed, store new cursor and apply change via glfw
+		rawCursor, isCustomCursor := fyneToNativeCursor(cursor)
 		w.cursor = cursor
-		if cursor == nil {
+
+		if rawCursor == nil {
 			viewport.SetInputMode(glfw.CursorMode, glfw.CursorHidden)
 		} else {
 			viewport.SetInputMode(glfw.CursorMode, glfw.CursorNormal)
-			viewport.SetCursor(cursor)
+			viewport.SetCursor(rawCursor)
+		}
+		if w.customCursor != nil {
+			w.customCursor.Destroy()
+			w.customCursor = nil
+		}
+		if isCustomCursor {
+			w.customCursor = rawCursor
 		}
 	}
 	if obj != nil && !w.objIsDragged(obj) {