+++ internal/driver/util_test.go
@@ -293,21 +293,53 @@ func TestReverseWalkVisibleObjectTree_Clip(t *testing.T) {
 	rect := canvas.NewRectangle(color.White)
 	rect.SetMinSize(fyne.NewSize(100, 100))
 	child := canvas.NewRectangle(color.Black)
-	base := fyne.NewContainerWithLayout(layout.NewGridLayout(1), rect, widget.NewScrollContainer(child))
+	base := fyne.NewContainerWithLayout(
+		layout.NewGridLayout(1),
+		rect,
+		widget.NewScrollContainer(child),
+		fyne.NewContainerWithLayout(
+			layout.NewGridLayout(2),
+			canvas.NewCircle(color.White),
+			canvas.NewCircle(color.White),
+			canvas.NewCircle(color.White),
+			&scrollable{},
+		),
+	)
 
-	clipPos := fyne.NewPos(0, 0)
-	clipSize := rect.MinSize()
+	var scClipPos, scrollableClipPos fyne.Position
+	var scClipSize, scrollableClipSize fyne.Size
 
 	driver.ReverseWalkVisibleObjectTree(base, func(object fyne.CanvasObject, position fyne.Position, clippingPos fyne.Position, clippingSize fyne.Size) bool {
 		if _, ok := object.(*widget.ScrollContainer); ok {
-			clipPos = clippingPos
-			clipSize = clippingSize
+			scClipPos = clippingPos
+			scClipSize = clippingSize
+		} else if _, ok = object.(fyne.Scrollable); ok {
+			scrollableClipPos = clippingPos
+			scrollableClipSize = clippingSize
 		}
 		return false
 	}, nil)
 
-	assert.Equal(t, fyne.NewPos(0, 104), clipPos)
-	assert.Equal(t, fyne.NewSize(100, 100), clipSize)
+	// layout:
+	// +-------------------------------+
+	// | 0,0: rect 100x100             |
+	// +-------------------------------+
+	// |            padding            |
+	// +-------------------------------+
+	// | 0,104: scroller 100x100       |
+	// +-------------------------------+
+	// |            padding            |
+	// +--------------+-+--------------+
+	// | circle 48x48 |p| circle 48x48 |
+	// +--------------+-+--------------+
+	// |            padding            |
+	// +--------------+-+--------------+
+	// | circle 48x48 |p| scrollable   |
+	// +--------------+-+--------------+
+	assert.Equal(t, fyne.NewPos(0, 104), scClipPos)
+	assert.Equal(t, fyne.NewSize(100, 100), scClipSize)
+	assert.Equal(t, fyne.NewPos(52, 260), scrollableClipPos)
+	assert.Equal(t, fyne.NewSize(48, 48), scrollableClipSize)
 }
 
 func TestWalkVisibleObjectTree(t *testing.T) {
@@ -331,21 +363,53 @@ func TestWalkVisibleObjectTree_Clip(t *testing.T) {
 	rect := canvas.NewRectangle(color.White)
 	rect.SetMinSize(fyne.NewSize(100, 100))
 	child := canvas.NewRectangle(color.Black)
-	base := fyne.NewContainerWithLayout(layout.NewGridLayout(1), rect, widget.NewScrollContainer(child))
+	base := fyne.NewContainerWithLayout(
+		layout.NewGridLayout(1),
+		rect,
+		widget.NewScrollContainer(child),
+		fyne.NewContainerWithLayout(
+			layout.NewGridLayout(2),
+			canvas.NewCircle(color.White),
+			canvas.NewCircle(color.White),
+			canvas.NewCircle(color.White),
+			&scrollable{},
+		),
+	)
 
-	clipPos := fyne.NewPos(0, 0)
-	clipSize := rect.MinSize()
+	var scClipPos, scrollableClipPos fyne.Position
+	var scClipSize, scrollableClipSize fyne.Size
 
 	driver.WalkVisibleObjectTree(base, func(object fyne.CanvasObject, position fyne.Position, clippingPos fyne.Position, clippingSize fyne.Size) bool {
 		if _, ok := object.(*widget.ScrollContainer); ok {
-			clipPos = clippingPos
-			clipSize = clippingSize
+			scClipPos = clippingPos
+			scClipSize = clippingSize
+		} else if _, ok = object.(fyne.Scrollable); ok {
+			scrollableClipPos = clippingPos
+			scrollableClipSize = clippingSize
 		}
 		return false
 	}, nil)
 
-	assert.Equal(t, fyne.NewPos(0, 104), clipPos)
-	assert.Equal(t, fyne.NewSize(100, 100), clipSize)
+	// layout:
+	// +-------------------------------+
+	// | 0,0: rect 100x100             |
+	// +-------------------------------+
+	// |            padding            |
+	// +-------------------------------+
+	// | 0,104: scroller 100x100       |
+	// +-------------------------------+
+	// |            padding            |
+	// +--------------+-+--------------+
+	// | circle 48x48 |p| circle 48x48 |
+	// +--------------+-+--------------+
+	// |            padding            |
+	// +--------------+-+--------------+
+	// | circle 48x48 |p| scrollable   |
+	// +--------------+-+--------------+
+	assert.Equal(t, fyne.NewPos(0, 104), scClipPos)
+	assert.Equal(t, fyne.NewSize(100, 100), scClipSize)
+	assert.Equal(t, fyne.NewPos(52, 260), scrollableClipPos)
+	assert.Equal(t, fyne.NewSize(48, 48), scrollableClipSize)
 }
 
 func TestWalkWholeObjectTree(t *testing.T) {
@@ -429,4 +493,52 @@ func (o objectTreeRenderer) MinSize() fyne.Size {
 }
 
 func (o objectTreeRenderer) Refresh() {
+}
+
+type scrollable struct {
+	pos  fyne.Position
+	size fyne.Size
+}
+
+var _ fyne.CanvasObject = (*scrollable)(nil)
+var _ fyne.Scrollable = (*scrollable)(nil)
+
+func (s *scrollable) Hide() {
+	panic("implement me")
+}
+
+func (s *scrollable) MinSize() fyne.Size {
+	return fyne.NewSize(1, 1)
+}
+
+func (s *scrollable) Move(position fyne.Position) {
+	s.pos = position
+}
+
+func (s *scrollable) Position() fyne.Position {
+	return s.pos
+}
+
+func (s *scrollable) Refresh() {
+	panic("implement me")
+}
+
+func (s *scrollable) Resize(size fyne.Size) {
+	s.size = size
+}
+
+func (s *scrollable) Scrolled(event *fyne.ScrollEvent) {
+	panic("implement me")
+}
+
+func (s *scrollable) Show() {
+	panic("implement me")
+}
+
+func (s *scrollable) Size() fyne.Size {
+	return s.size
+}
+
+func (s *scrollable) Visible() bool {
+	return true
 }