+++ CHANGELOG.md
@@ -13,6 +13,7 @@ These changes likely break some apps, please read the
 * The second parameter of `theme.NewThemedResource` was removed, it was previously ignored
 * `fyne.Theme` moved to `fyne.LegacyTheme` and can be load to a new theme using `theme.FromLegacy`
 * A new, more flexible, Theme interface has been created that we encourage developers to use
+* The desktop.Cursor definition was renamed desktop.StandardCursor to make way for custom cursors
 
 ### Added